requirements.txt

fastapi
uvicorn
tree_sitter
flake8
bandit
chromadb
python-dotenv
